🚚 Rename : 컴포넌트 위치 변경 (Tip, Widget관련)
